requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
